fix NodeBalancerConfigFields dep (11)
