+++ src/services/nodebalancers/configs.ts
@@ -8,28 +8,6 @@ import { combineConfigNodeAddressAndPort } from './utils';
 
 type Page<T> = Linode.ResourcePage<T>;
 
-/** DEPRECATED!!!! */
-/** DEPRECATED!!!! */
-/** DEPRECATED!!!! */
-export interface NodeBalancerConfigFields {
-  id?: number;
-  algorithm?: 'roundrobin' | 'leastconn' | 'source';
-  check_attempts?: number /** 1..30 */;
-  check_body?: string;
-  check_interval?: number;
-  check_passive?: boolean;
-  check_path?: string;
-  check_timeout?: number /** 1..30 */;
-  check?: 'none' | 'connection' | 'http' | 'http_body';
-  cipher_suite?: 'recommended' | 'legacy';
-  port?: number /** 1..65535 */;
-  protocol?: 'http' | 'https' | 'tcp';
-  ssl_cert?: string;
-  ssl_key?: string;
-  stickiness?: 'none' | 'table' | 'http_cookie';
-  nodes: Linode.NodeBalancerConfigNodeFields[];
-}
-
 export interface CreateNodeBalancerConfig {
   port?: number;
   protocol?: 'http' | 'https' | 'tcp';
