+++ src/services/nodebalancers/index.ts
@@ -3,3 +3,5 @@ export * from './nodebalancers';
 export * from './configs';
 
 export * from './configNodes';
+
+export * from './interfaces';
